# packages/sdk-java/managed-agent-server/src/main/resources/openapi/managed-agent-public-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qwen Managed Agent Public and WebShell API",
    "version": "1.26.0",
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "trustedTenant": [],
      "trustedActor": []
    }
  ],
...
  "paths": {
...
    "/v1/agents/{agentId}": {
      "parameters": [
        {
          "name": "agentId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      ],
      "get": {
        "tags": ["Public Resources"],
        "operationId": "getAgent",
        "x-qwen-implementation-status": "planned",
        "parameters": [
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested or current agent revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinition"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Public Resources"],
        "operationId": "updateAgent",
        "x-qwen-implementation-status": "planned",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDefinitionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "New immutable revision published for this agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinition"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/agents/sessions": {
      "post": {
        "tags": ["Public Sessions"],
        "operationId": "createSession",
        "x-qwen-implementation-status": "implemented",
        "description": "Workspace selection is persisted with the original actor-scoped creation receipt. With the deployment's G0 file admission opt-in, qwen-code Sessions in an authorized, preregistered Workspace may include initial input under the fixed preapproved file-tool profile. Public callers cannot select the profile. Otherwise Workspace creation must have empty input. Later Workspace submit, cancel and lifecycle operations remain gated. Check replay before resolving defaults; a durable binding is not proof of Runtime readiness.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Session accepted after durable command admission.",
            "headers": {
              "X-Qwen-Idempotent-Replay": {
                "$ref": "#/components/headers/IdempotentReplay"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSession"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "get": {
        "tags": ["Public Sessions"],
        "operationId": "listSessions",
        "x-qwen-implementation-status": "implemented",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ListLimit"
          },
          {
            "name": "include_archived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-qwen-implementation-status": "planned"
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions ordered by updated_at and id descending.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSessionList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/agents/sessions/{sessionId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "get": {
        "tags": ["Public Sessions"],
        "operationId": "getSession",
        "x-qwen-implementation-status": "implemented",
        "responses": {
          "200": {
            "description": "Session state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSession"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Public Sessions"],
        "operationId": "updateSession",
        "x-qwen-implementation-status": "partial",
        "description": "Renames the Session. Shipped before the v1.10 lifecycle review, which does not cover it.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Renamed Session.",
            "headers": {
              "X-Qwen-Idempotent-Replay": {
                "$ref": "#/components/headers/IdempotentReplay"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSession"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": ["Public Sessions"],
        "operationId": "deleteSession",
        "x-qwen-implementation-status": "implemented",
        "description": "Accepts an active, closed or archived Session. 409 turn_active while a Turn is active, session_operation_active while another lifecycle change is pending, and workspace_unavailable for a Workspace-bound Session. The same key from the same actor returns the original operation, also after the Session was deleted.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted. The Session stops accepting input and reads deleting. If it was active, the operation first closes it as a close does; it then drains the Session's Runtime binding and leaves a tombstone: Session reads return 404, while its operations stay readable. The shared Workspace is never deleted, and no Session content is erased yet. 202 does not mean that tools have stopped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCommandOperation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/agents/sessions/{sessionId}/events": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "post": {
        "tags": ["Public Sessions"],
        "operationId": "postSessionEvent",
        "x-qwen-implementation-status": "partial",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionEventRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Input or cancellation durably accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAdmission"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "tags": ["Public Sessions"],
        "operationId": "getSessionEvents",
        "x-qwen-implementation-status": "implemented",
        "description": "Returns JSON by default. Accept text/event-stream or stream=true to receive SSE. Last-Event-ID takes precedence over after. A cursor below replay_floor_sequence gets 409 cursor_expired as JSON; a stream instead sends one agent.session.resync_required frame and closes.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AfterSequence"
          },
          {
            "$ref": "#/components/parameters/EventLimit"
          },
          {
            "name": "stream",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event page or SSE stream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEventList"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE frames whose id is the public sequence and whose data is a PublicEvent. An expired cursor ends the stream with one agent.session.resync_required frame that has no id; x-qwen-resync-frame is its data."
                },
                "x-qwen-resync-frame": {
                  "$ref": "#/components/schemas/SessionResyncRequired"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/CursorExpired"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/agents/sessions/{sessionId}/turns": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "get": {
        "tags": ["Public Resources"],
        "operationId": "listTurns",
        "x-qwen-implementation-status": "implemented",
        "description": "Turns of this Session, newest first: by creation time in milliseconds, which created_at shows in whole seconds, then by id, both descending. next_cursor names the last Turn of the page and is null on the last page; a page never repeats or skips a Turn that existed when the first page was read. 400 invalid_limit for a 32-bit integer limit outside 1 to 100, invalid_request for a limit that is not a 32-bit integer, and invalid_cursor for a cursor that does not decode to the createdAt:turnId form this server issues. 404 session_not_found for an unknown, deleted or unreadable Session.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ListLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Turn page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTurnList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/agents/sessions/{sessionId}/turns/{turnId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        },
        {
          "$ref": "#/components/parameters/TurnId"
        }
      ],
      "get": {
        "tags": ["Public Resources"],
        "operationId": "getTurn",
        "x-qwen-implementation-status": "implemented",
        "description": "404 session_not_found for an unknown, deleted or unreadable Session, and turn_not_found for a Turn that this Session does not have; an id that differs from a Turn's only by trailing spaces does not find it. 400 invalid_request for a Turn id longer than 64 characters.",
        "responses": {
          "200": {
            "description": "Turn state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTurn"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/agents/sessions/{sessionId}/items": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "get": {
        "tags": ["Public Resources"],
        "operationId": "listItems",
        "x-qwen-implementation-status": "partial",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "$ref": "#/components/parameters/ListLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Materialized items and snapshot watermark.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicItemList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/v1/agents/sessions/{sessionId}/operations/{operationId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        },
        {
          "name": "operationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        }
      ],
      "get": {
        "tags": ["Public Sessions"],
        "operationId": "getSessionCwdOperation",
        "x-qwen-implementation-status": "implemented",
        "description": "Reads one operation of the Session. Close, archive and delete operations are served now; cwd, command and task operations arrive with their slices. 404 operation_not_found for an unknown operation.",
        "responses": {
          "200": {
            "description": "Authorized operation with its latest durable status, readable without Runtime startup and after the Session was deleted. Reading is not a replay, so replayed is false.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicOperation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/v1/agents/sessions/{sessionId}/close": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "post": {
        "tags": ["Public Sessions"],
        "operationId": "closeSession",
        "x-qwen-implementation-status": "implemented",
        "description": "Accepts an active Session; 409 session_state_conflict otherwise. 409 turn_active while a Turn is active, session_operation_active while another lifecycle change is pending, and workspace_unavailable for a Workspace-bound Session.",
        "responses": {
          "202": {
            "description": "Close accepted. The Session stops accepting input and reads closing. The operation asks the Hosted Harness to close the Session and completes once no Harness holds the Session's journal writer under an unexpired lease (the Harness that holds the Session seals it when closing) and the Session's Runtime binding drained; the Session then reads closed. 202 does not mean that tools have stopped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCommandOperation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "trustedTenant": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Qwen-Tenant-Id",
        "description": "Injected by the trusted authenticated gateway, never trusted from browser input."
      },
      "trustedActor": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Qwen-Actor-Id",
        "x-qwen-implementation-status": "planned",
        "description": "Verified actor injected by the authenticated gateway together with tenant; browser-supplied headers are stripped. Resource ACLs and Action responder eligibility are mandatory. This header is not a browser credential."
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
    },
...
    "responses": {
      "Forbidden": {
        "description": "The authenticated actor belongs to another tenant than X-Qwen-Tenant-Id or has an invalid ID (actor_scope_mismatch, which the tenant filter answers on every /v1/agents/ and WebShell route), or the resource is readable but the actor lacks this operation or the original Action responder permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource absent or outside the caller tenant scope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AgentDefinition": {
        "type": "object",
        "required": ["id", "object", "revision", "digest", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128
          },
          "object": {
            "const": "agent"
          },
          "revision": {
            "type": "string",
            "maxLength": 128
          },
          "digest": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}